fs/drivers/fat: Use the same iterate handler
